+++ apps/backoffice-v2/src/domains/individuals/fetchers.ts
@@ -8,11 +8,19 @@ import { HitSchema } from '@/lib/blocks/components/AmlBlock/utils/aml-adapter';
 
 import { apiClient } from '@/common/api-client/api-client';
 
+export const EntityType = {
+  BUSINESS: 'business',
+  UBO: 'ubo',
+  DIRECTOR: 'director',
+} as const;
+
+export const EndUserVariantSchema = z.enum([EntityType.UBO, EntityType.DIRECTOR]);
+
 export const EndUserSchema = z.object({
   id: z.string(),
   firstName: z.string(),
   lastName: z.string(),
-  email: z.string().optional(),
+  email: z.string().nullable().optional(),
   gender: z.string().nullable(),
   nationality: z.string().nullable(),
   address: z.string().nullable(),
@@ -38,6 +46,8 @@ export const EndUserSchema = z.object({
       }),
     })
     .optional(),
+  variant: z.enum(['director', 'ubo']).optional().nullable(),
+  createdFrom: z.enum(['user', 'analyst', 'registry']).nullable().optional(),
 });
 
 export const EndUsersSchema = z.array(EndUserSchema);