+++ services/workflows-service/src/common/guards/token-guard/token-auth.module.ts
@@ -1,7 +1,7 @@
 import { WorkflowTokenRepository } from '@/auth/workflow-token/workflow-token.repository';
 import { WorkflowTokenService } from '@/auth/workflow-token/workflow-token.service';
 import { WorkflowAuthGuard } from '@/common/guards/workflow-guard/workflow-auth.guard';
-import { Module } from '@nestjs/common';
+import { Module, forwardRef } from '@nestjs/common';
 import { CustomerService } from '@/customer/customer.service';
 import { UiDefinitionService } from '@/ui-definition/ui-definition.service';
 import { CustomerRepository } from '@/customer/customer.repository';
@@ -34,14 +34,15 @@ import { PasswordService } from '@/auth/password/password.service';
 import { SalesforceIntegrationRepository } from '@/salesforce/salesforce-integration.repository';
 import { NotionService } from '@/notion/notion.service';
 import { FileRepository } from '@/storage/storage.repository';
-import { WebhookHttpService } from '@/alert/webhook-manager/webhook-manager.service';
 import { HttpModule, HttpService } from '@nestjs/axios';
 import { WorkflowRuntimeDataActorService } from '@/workflow/workflow-runtime-data-actor.service';
 import { AssessmentsService } from '@/assessments/assessments.service';
 import { UnifiedApiClient } from '@/common/utils/unified-api-client/unified-api-client';
+// eslint-disable-next-line import/no-cycle
+import { KycModule } from '@/kyc/kyc.module';
 
 @Module({
-  imports: [HttpModule],
+  imports: [HttpModule, forwardRef(() => KycModule)],
   providers: [
     MerchantMonitoringClient,
     WorkflowTokenRepository,
@@ -67,10 +68,6 @@ import { UnifiedApiClient } from '@/common/utils/unified-api-client/unified-api-
     FileRepository,
     UserService,
     UserRepository,
-    {
-      provide: WebhookHttpService,
-      useExisting: HttpService,
-    },
     SalesforceService,
     SalesforceIntegrationRepository,
     RiskRuleService,