+++ services/workflows-service/src/webhooks/webhooks.module.ts
@@ -1,86 +1,43 @@
-import { AuthModule } from '@/auth/auth.module';
-import { ACLModule } from '@/common/access-control/acl.module';
-import { CustomerModule } from '@/customer/customer.module';
-import { PrismaModule } from '@/prisma/prisma.module';
-import { ProjectModule } from '@/project/project.module';
-import { WorkflowDefinitionModule } from '@/workflow-defintion/workflow-definition.module';
+import { createBullBoard } from '@bull-board/api';
+import { ExpressAdapter } from '@bull-board/express';
 import { HttpModule } from '@nestjs/axios';
-import { forwardRef, Module } from '@nestjs/common';
-import { WorkflowRuntimeDataRepository } from '@/workflow/workflow-runtime-data.repository';
-import { EndUserRepository } from '@/end-user/end-user.repository';
-import { EndUserService } from '@/end-user/end-user.service';
-import { BusinessRepository } from '@/business/business.repository';
-import { EntityRepository } from '@/common/entity/entity.repository';
-import { FileService } from '@/providers/file/file.service';
-import { WorkflowEventEmitterService } from '@/workflow/workflow-event-emitter.service';
-import { UserService } from '@/user/user.service';
-import { SalesforceService } from '@/salesforce/salesforce.service';
-import { WorkflowTokenService } from '@/auth/workflow-token/workflow-token.service';
-import { UiDefinitionService } from '@/ui-definition/ui-definition.service';
-import { StorageService } from '@/storage/storage.service';
-import { UserRepository } from '@/user/user.repository';
-import { SalesforceIntegrationRepository } from '@/salesforce/salesforce-integration.repository';
-import { WorkflowTokenRepository } from '@/auth/workflow-token/workflow-token.repository';
-import { UiDefinitionRepository } from '@/ui-definition/ui-definition.repository';
-import { FileRepository } from '@/storage/storage.repository';
-import { HookCallbackHandlerService } from '@/workflow/hook-callback-handler.service';
-import { FilterService } from '@/filter/filter.service';
-import { FilterRepository } from '@/filter/filter.repository';
-import { WebhooksController } from '@/webhooks/webhooks.controller';
-import { WebhooksService } from '@/webhooks/webhooks.service';
-import { BusinessService } from '@/business/business.service';
-import { BusinessReportModule } from '@/business-report/business-report.module';
-import { AlertModule } from '@/alert/alert.module';
-import { DataAnalyticsModule } from '@/data-analytics/data-analytics.module';
-import { AlertDefinitionModule } from '@/alert-definition/alert-definition.module';
-import { RuleEngineModule } from '@/rule-engine/rule-engine.module';
-import { SentryService } from '@/sentry/sentry.service';
-import { WorkflowModule } from '@/workflow/workflow.module';
-import { DocumentModule } from '@/document/document.module';
+import { Inject, MiddlewareConsumer, Module } from '@nestjs/common';
+
+import { AppLoggerModule } from '@/common/app-logger/app-logger.module';
+import { QueueModule } from '@/common/queue/queue.module';
+import { MonitoringModule } from '@/common/monitoring/monitoring.module';
+import { BULLBOARD_INSTANCE_INJECTION_TOKEN } from '@/common/queue/types';
+import type { BullBoardInjectedInstance } from '@/common/queue/types';
+import { BullBoardAuthMiddleware } from './bull-board.auth.middleware';
+import { WebhooksService } from './webhooks.service';
 
 @Module({
-  controllers: [WebhooksController],
-  imports: [
-    ACLModule,
-    forwardRef(() => AuthModule),
-    HttpModule,
-    ProjectModule,
-    PrismaModule,
-    CustomerModule,
-    BusinessReportModule,
-    WorkflowDefinitionModule,
-    AlertModule,
-    DataAnalyticsModule,
-    AlertDefinitionModule,
-    RuleEngineModule,
-    WorkflowModule,
-    DocumentModule,
-  ],
+  imports: [AppLoggerModule, HttpModule, QueueModule, MonitoringModule],
   providers: [
-    WorkflowRuntimeDataRepository,
-    EndUserService,
-    EndUserRepository,
-    BusinessService,
-    BusinessRepository,
-    EntityRepository,
-    FileService,
-    FileRepository,
-    WorkflowEventEmitterService,
-    UserService,
-    UserRepository,
-    SalesforceService,
-    SalesforceIntegrationRepository,
-    WorkflowTokenService,
-    WorkflowTokenRepository,
-    UiDefinitionService,
-    UiDefinitionRepository,
-    StorageService,
-    HookCallbackHandlerService,
-    FilterService,
-    FilterRepository,
     WebhooksService,
-    SentryService,
+    {
+      provide: BULLBOARD_INSTANCE_INJECTION_TOKEN,
+      useFactory: (): BullBoardInjectedInstance => {
+        const serverAdapter = new ExpressAdapter();
+        serverAdapter.setBasePath('/api/queues');
+
+        const boardInstance = createBullBoard({ queues: [], serverAdapter });
+
+        return { boardInstance, serverAdapter };
+      },
+    },
   ],
-  exports: [],
+  exports: [WebhooksService, BULLBOARD_INSTANCE_INJECTION_TOKEN],
 })
-export class WebhooksModule {}
+export class WebhooksModule {
+  constructor(
+    @Inject(BULLBOARD_INSTANCE_INJECTION_TOKEN)
+    private bullBoard: BullBoardInjectedInstance,
+  ) {}
+
+  configure(consumer: MiddlewareConsumer) {
+    consumer
+      .apply(BullBoardAuthMiddleware, this.bullBoard.serverAdapter.getRouter())
+      .forRoutes('/api/queues');
+  }
+}