+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @ballerine/workflow-core
 
+## 0.6.142
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.118
+
+## 0.6.141
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.117
+
+## 0.6.140
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.116
+
+## 0.6.139
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.115
+
+## 0.6.138
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.114
+
+## 0.6.137
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.113
+
+## 0.6.136
+
+### Patch Changes
+
+- version bump
+
 ## 0.6.135
 
 ### Patch Changes